Fix grid-rendering issue
Event blocks on BlockTracks in the EventsGrid had the tendency to be
slightly mis-positioned, by 1px. In isolation this isn't noticeable,
but with tight clusters of notes there were odd, irregular gaps. This
is because I was using `left`, which doesn't do sub-pixel rounding well.

Switched to Transform and it's SO NICE AND UNIFORM NOW.